Revert "Big update"
