--- deno.json
@@ -1,6 +1,0 @@
-{
-  "imports": {
-    "@std/assert": "jsr:@std/assert@^1.0.11",
-    "@std/cli": "jsr:@std/cli@^1.0.13"
-  }
-}
