--- deno.lock
@@ -1,28 +1,0 @@
-{
-  "version": "4",
-  "specifiers": {
-    "jsr:@std/assert@^1.0.11": "1.0.11",
-    "jsr:@std/cli@^1.0.13": "1.0.13",
-    "jsr:@std/internal@^1.0.5": "1.0.5"
-  },
-  "jsr": {
-    "@std/assert@1.0.11": {
-      "integrity": "2461ef3c368fe88bc60e186e7744a93112f16fd110022e113a0849e94d1c83c1",
-      "dependencies": [
-        "jsr:@std/internal"
-      ]
-    },
-    "@std/cli@1.0.13": {
-      "integrity": "5db2d95ab2dca3bca9fb6ad3c19908c314e93d6391c8b026725e4892d4615a69"
-    },
-    "@std/internal@1.0.5": {
-      "integrity": "54a546004f769c1ac9e025abd15a76b6671ddc9687e2313b67376125650dc7ba"
-    }
-  },
-  "workspace": {
-    "dependencies": [
-      "jsr:@std/assert@^1.0.11",
-      "jsr:@std/cli@^1.0.13"
-    ]
-  }
-}
