+++ examples/fib.fscript
