+++ examples/functions.fscript
