+++ examples/pipe.fscript
