+++ examples/recursive.fscript
