--- examples/var_declaration.farpy
@@ -1,3 +1,0 @@
-new mut x: int = 9
-x++ ** 2 // x = 100
-x-- // x = 99
