+++ examples/var_declaration.fscript
@@ -1,0 +1,3 @@
+new mut x: int = 9
+x++**2 // x = 100
+x--
